updated epd-fuse installation
updated epd-fuse installation. This has been updated in our code at https://github.com/PiSupply/PaPiRus but not in the readme. Apologies for that.

You need the panel version and other parameters after the make commands for this to work. Hopefully this should work better with the latest gratis epd-fuse code.